mobile/android/components/extensions/test/mochitest/test_ext_pageAction.html

<!DOCTYPE HTML>
<html>
<head>
  <title>PageAction Test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/contents/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

function backgroundScript() {
  browser.test.assertTrue("pageAction" in browser, "Namespace 'pageAction' exists in browser");
  browser.test.assertTrue("show" in browser.pageAction, "API method 'show' exists in browser.pageAction");

  // TODO: Use the Tabs API to obtain the tab ids for showing pageActions.
  let tabId = 1;

  browser.pageAction.show(tabId);
  browser.test.sendMessage("page-action-shown");

  browser.test.notifyPass("page-action");
}

add_task(function* test_contentscript() {
  let extension = ExtensionTestUtils.loadExtension({
    background: "(" + backgroundScript.toString() + ")()",
    manifest: {
      "name": "PageAction Extension",
      "page_action": {
        "default_title": "Page Action",
      },
    },
  });

  yield extension.startup();
  yield extension.awaitMessage("page-action-shown");

  is(isPageActionShown(extension.id), true, "The PageAction should be shown");

  yield extension.awaitFinish("page-action");
  yield extension.unload();

  is(isPageActionShown(extension.id), false, "The PageAction should be removed after unload");
});
</script>

</body>
</html>
